Fixed hive.exec.mode.local.auto=true path resolution
The classpath for jars in local mode points to a real file on disk. However, the JobSubmitter was cutting off the protocol part of the path. By default DistributedCache assumes that protocol-less files come from hdfs.
This patch passes the entire path to the DistributedCache, which allowed DistributedCache to find it in the file system.