AE-321: Fixed local file resolution in local mode.
In local mode, hadoop was not able to resolve local files.
It was assuming those files are on hdfs. The root cause was
that the path passed to the DistributedCache by the
JobSubmitter did not include the protocol. This change adds
back the protocol, and allows the resolution to happen correctly.